@@ -21,74 +21,74 @@
     "tsc": "tsc -p tsconfig.json"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.9",
-    "@fortawesome/free-solid-svg-icons": "^5.6.0",
-    "@fortawesome/react-fontawesome": "^0.1.3",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
+    "@fortawesome/react-fontawesome": "^0.1.4",
     "@types/history": "^4.7.2",
-    "@types/node": "^11.10.5",
-    "@types/query-string": "^6.2.0",
-    "@types/react": "^16.8.7",
+    "@types/node": "^12.0.7",
+    "@types/query-string": "^6.3.0",
+    "@types/react": "^16.8.19",
     "@types/react-bootstrap": "^0.32.17",
     "@types/react-bootstrap-table": "^4.3.7",
-    "@types/react-dom": "^16.8.2",
-    "@types/react-redux": "^7.0.3",
-    "@types/react-router": "^4.4.5",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-redux": "^7.0.9",
+    "@types/react-router": "^5.0.1",
     "@types/react-router-bootstrap": "^0.24.5",
-    "@types/react-router-dom": "^4.3.1",
+    "@types/react-router-dom": "^4.3.3",
     "@types/react-router-redux": "^5.0.18",
-    "@types/react-select": "^2.0.17",
+    "@types/react-select": "^2.0.19",
     "bootstrap": "^3.4.0",
-    "history": "^4.7.2",
+    "history": "^4.9.0",
     "hoist-non-react-statics": "^3.3.0",
     "query-string": "^6.3.0",
-    "react": "^16.8.4",
+    "react": "^16.8.6",
     "react-bootstrap": "^0.32.4",
     "react-bootstrap-table": "^4.3.1",
-    "react-dom": "^16.8.4",
-    "react-is": "^16.8.4",
-    "react-redux": "^6.0.1",
-    "react-router": "^4.3.1",
-    "react-router-bootstrap": "^0.24.4",
-    "react-router-dom": "^4.3.1",
+    "react-dom": "^16.8.6",
+    "react-is": "^16.8.6",
+    "react-redux": "^7.1.0",
+    "react-router": "^5.0.1",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.0.1",
     "react-router-redux": "^4.0.8",
-    "react-select": "^2.4.3",
+    "react-select": "^3.0.4",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
-    "ts-node": "^8.0.3",
-    "typesafe-actions": "^3.2.1",
-    "typescript": "^3.3.3333",
-    "universal-cookie": "^3.0.7",
-    "utility-types": "^3.4.1"
+    "ts-node": "^8.2.0",
+    "typesafe-actions": "^4.4.0",
+    "typescript": "^3.5.1",
+    "universal-cookie": "^4.0.0",
+    "utility-types": "^3.7.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.0",
-    "@babel/plugin-proposal-class-properties": "^7.2.1",
-    "@babel/plugin-proposal-object-rest-spread": "^7.2.0",
-    "@babel/preset-env": "^7.2.0",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.1.0",
-    "babel-loader": "^8.0.2",
-    "clean-webpack-plugin": "^0.1.19",
-    "css-loader": "^0.28.11",
+    "@babel/preset-typescript": "^7.3.3",
+    "babel-loader": "^8.0.6",
+    "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^3.0.0",
     "html-webpack-plugin": "^3.2.0",
-    "image-webpack-loader": "^4.6.0",
-    "mini-css-extract-plugin": "^0.4.3",
-    "node-sass": "^4.11.0",
+    "image-webpack-loader": "^5.0.0",
+    "mini-css-extract-plugin": "^0.7.0",
+    "node-sass": "^4.12.0",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "sass-loader": "^7.1.0",
     "source-map": "^0.7.3",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.21.0",
+    "style-loader": "^0.23.1",
     "svg-url-loader": "^2.3.2",
-    "terser": "^3.16.1",
-    "tslint": "^5.11.0",
-    "tslint-react": "^3.6.0",
-    "uglifyjs-webpack-plugin": "^2.0.1",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-middleware": "^3.4.0",
-    "webpack-merge": "^4.1.4"
+    "terser": "^4.0.0",
+    "tslint": "^5.17.0",
+    "tslint-react": "^4.0.0",
+    "uglifyjs-webpack-plugin": "^2.1.3",
+    "url-loader": "^2.0.0",
+    "webpack": "^4.33.0",
+    "webpack-cli": "^3.3.4",
+    "webpack-dev-middleware": "^3.7.0",
+    "webpack-merge": "^4.2.1"
   }
 }
