@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using Database.Models.Addressing;
 using Microsoft.EntityFrameworkCore;
 
@@ -7,9 +10,12 @@ namespace Database.Repositories.Implementation
     {
         private readonly DbSet<Street> _streets;
 
-        public StreetsRepository(DbSet<Street> streets)
-        {
-            _streets = streets;
-        }
+        public StreetsRepository(DbSet<Street> streets) => _streets = streets;
+
+        public IEnumerable<Street> GetLocalityId(Guid localityId) => _streets.Where(locality => locality.LocalityId == localityId);
+
+        public void Add(Street street) => _streets.Add(street);
+
+        public bool Exists(Guid localityId, string name) => _streets.Any(street => street.LocalityId == localityId && street.Name == name);
     }
 }
