@@ -1,3 +1,6 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using Database.Models.Addressing;
 using Microsoft.EntityFrameworkCore;
 
@@ -7,9 +10,14 @@ namespace Database.Repositories.Implementation
     {
         private readonly DbSet<Territory> _territories;
 
-        public TerritoriesRepository(DbSet<Territory> territories)
-        {
-            _territories = territories;
-        }
+        public TerritoriesRepository(DbSet<Territory> territories) => _territories = territories;
+
+        public IEnumerable<Territory> Get() => _territories;
+
+        public IEnumerable<Territory> GetByParentId(Guid parentId) => _territories.Where(locality => locality.ParentId == parentId);
+
+        public void Add(Territory territory) => _territories.Add(territory);
+
+        public bool Exists(Guid parentId, string name) => _territories.Any(territory => territory.ParentId == parentId && territory.Name == name);
     }
 }
