@@ -1,7 +1,16 @@
 import ApiBase from "./api-base";
 import {HttpMethod} from "../constants/http-method";
-import UserModel from "./models/user-model";
-import AuthData from "./models/auth-data";
+import {UserStatus} from "../constants/user-status";
+
+export interface UserModel {
+    readonly id: string;
+    readonly email: string;
+    readonly status: UserStatus;
+    readonly firstName: string;
+    readonly lastName: string;
+    readonly patronymic: string;
+    readonly phoneNumber: string;
+}
 
 export default class UsersApi extends ApiBase {
 
@@ -9,19 +18,6 @@ export default class UsersApi extends ApiBase {
         return fetch(`${this.SERVER_HOST}/api/users?start=${start}&size=${size}`, {
             method: HttpMethod.Get,
             headers: this.authorizedHeaders()
-        })
-            .then((response: Response) => {
-                if (response.ok) {
-                    return response.json();
-                }
-            });
-    }
-
-    public static changePassword(password: string, token?: string): Promise<AuthData> {
-        return fetch(`${this.SERVER_HOST}/api/users/change-password`, {
-            method: HttpMethod.Put,
-            headers: this.authorizedHeaders(token),
-            body: JSON.stringify(password)
         })
             .then((response: Response) => {
                 if (response.ok) {
