@@ -3,16 +3,7 @@ import React, {FunctionComponent} from "react";
 import {Route, Router, Switch} from "react-router-dom";
 import history from "./store/history";
 import {AuthRoute, NavigationBar} from "./components";
-import {
-    Auth,
-    Cabinet,
-    ChangePassword,
-    Error404,
-    Home,
-    Invite,
-    Register,
-    RestorePassword
-} from "./pages";
+import {Auth, Cabinet, ChangePassword, Error404, Home, Invite, Register, RestorePassword} from "./pages";
 
 export const App: FunctionComponent = () => (
     <Router history={history}>
