@@ -3,11 +3,11 @@ import React, {Component, ReactNode} from "react";
 import {connect} from "react-redux";
 import {RootState} from "../../store";
 import FormGroup from "react-bootstrap/lib/FormGroup";
-import ControlLabel from "react-bootstrap/lib/ControlLabel";
 import FormControl from "react-bootstrap/lib/FormControl";
 import Col from "react-bootstrap/lib/Col";
 import Button from "react-bootstrap/lib/Button";
 import Form from "react-bootstrap/lib/Form";
+import ControlLabel from "react-bootstrap/lib/ControlLabel";
 import Select from "react-select";
 import {roleNameDetails} from "../../constants/role-name";
 import InvitesApi from "../../api/invites-api";
