@@ -1,19 +1,8 @@
-import {Auth} from "./auth";
-import {Cabinet} from "./cabinet";
-import {ChangePassword} from "./change-password";
-import {Error404} from "./error404";
-import {Home} from "./home";
-import {Invite} from "./invite";
-import {Register} from "./register";
-import {RestorePassword} from "./restore-password";
-
-export {
-    Auth,
-    Cabinet,
-    ChangePassword,
-    Error404,
-    Home,
-    Invite,
-    Register,
-    RestorePassword
-};
+export {Auth} from "./auth";
+export {Cabinet} from "./cabinet";
+export {ChangePassword} from "./change-password";
+export {Error404} from "./error404";
+export {Home} from "./home";
+export {Invite} from "./invite";
+export {Register} from "./register";
+export {RestorePassword} from "./restore-password";
